requirements.txt
django
djangorestframework
django-filter
psycopg2-binary
Pillow
